Remove repeated method
