+++ beep/protocol/maccor.py
@@ -13,11 +13,13 @@ import csv
 from copy import deepcopy
 
 import pandas as pd
+import numpy as np
 import xmltodict
 from docopt import docopt
 from monty.serialization import loadfn
 from beep import logger, __version__
 from beep.protocol import PROCEDURE_TEMPLATE_DIR
+from beep.conversion_schemas import MACCOR_WAVEFORM_CONFIG
 from beep.utils import KinesisEvents, DashOrderedDict
 s = {'service': 'ProtocolGenerator'}
 
@@ -864,154 +866,63 @@ class Procedure(DashOrderedDict):
         return self
 
 
-def generate_protocol_files_from_csv(csv_filename, output_directory):
+def generate_maccor_waveform_file(df, file_prefix, file_directory, mwf_config=None):
     """
-    Generates a set of protocol files from csv filename input by
-    reading protocol file input corresponding to each line of
-    the csv file. Writes a csv file that.
+    Helper function that takes in a variable power waveform and outputs a maccor waveform file (.MWF), which is read by
+    the cycler when the procedure file has a "fast-wave" step.
+    Relevant parameters to generate the .mwf files governing the input mode, charge/discharge limits, end conditions
+    and scaling are stored in /conversion_schemas/maccor_waveform_conversion.yaml
 
     Args:
-        csv_filename (str): CSV containing protocol file parameters.
-        output_directory (str): directory in which to place the output files
-    """
-    # Read csv file
-    protocol_params_df = pd.read_csv(csv_filename)
-
-    new_files = []
-    names = []
-    result = ''
-    message = {'comment': '',
-               'error': ''}
-    for index, protocol_params in protocol_params_df.iterrows():
-        template = protocol_params['template']
-
-        # Switch for template invocation
-        if template == "EXP.000":
-            procedure = Procedure.from_exp(
-                **protocol_params[["cutoff_voltage", "charge_rate", "discharge_rate"]]
-            )
-        elif template == 'diagnosticV2.000':
-            diag_params_df = pd.read_csv(os.path.join(PROCEDURE_TEMPLATE_DIR,
-                                                      "PreDiag_parameters - DP.csv"))
-            diagnostic_params = diag_params_df[diag_params_df['diagnostic_parameter_set'] ==
-                                               protocol_params['diagnostic_parameter_set']].squeeze()
-
-            # TODO: should these be separated?
-            procedure = Procedure.from_regcyclev2(
-                protocol_params
-            )
-            procedure.add_procedure_diagcyclev2(
-                protocol_params["capacity_nominal"], diagnostic_params
-            )
-
-        # TODO: how are these different?
-        elif template in ['diagnosticV3.000', 'diagnosticV4.000']:
-            diag_params_df = pd.read_csv(os.path.join(PROCEDURE_TEMPLATE_DIR,
-                                                      "PreDiag_parameters - DP.csv"))
-            diagnostic_params = diag_params_df[diag_params_df['diagnostic_parameter_set'] ==
-                                               protocol_params['diagnostic_parameter_set']].squeeze()
-
-            procedure = Procedure.generate_procedure_regcyclev3(index, protocol_params)
-            procedure.generate_procedure_diagcyclev3(
-                    protocol_params["capacity_nominal"], diagnostic_params
-            )
-        else:
-            warnings.warn("Unsupported file template {}, skipping.".format(template))
-            result = "error"
-            message = {'comment': 'Unable to find template: ' + template,
-                       'error': 'Not Found'}
-            continue
-
-        filename_prefix = '_'.join(
-            [protocol_params["project_name"], '{:06d}'.format(protocol_params["seq_num"])])
-        filename = "{}.000".format(filename_prefix)
-        filename = os.path.join(output_directory, 'procedures', filename)
-        logger.info(filename, extra=s)
-        if not os.path.isfile(filename):
-            procedure.to_file(filename)
-            new_files.append(filename)
-            names.append(filename_prefix + '_')
-
-        elif '.sdu' in template:
-            logger.warning('Schedule file generation not yet implemented', extra=s)
-            result = "error"
-            message = {'comment': 'Schedule file generation is not yet implemented',
-                       'error': 'Not Implemented'}
-
-    # This block of code produces the file containing all of the run file
-    # names produced in this function call. This is to make starting tests easier
-    _, namefile = os.path.split(csv_filename)
-    namefile = namefile.split('_')[0] + '_names_'
-    namefile = namefile + datetime.datetime.now().strftime("%Y%m%d_%H%M") + '.csv'
-    with open(os.path.join(output_directory, "names", namefile), 'w', newline='') as outputfile:
-        wr = csv.writer(outputfile)
-        for name in names:
-            wr.writerow([name])
-    outputfile.close()
-
-    if not result:
-        result = "success"
-        message = {'comment': 'Generated {} protocols'.format(str(len(new_files))),
-                   'error': ''}
-
-    return new_files, result, message
-
-
-def process_csv_file_list_from_json(
-        file_list_json,
-        processed_dir='data-share/protocols/'):
-    """
-
-    Args:
-        file_list_json (str):
-        processed_dir (str):
+        df (pd.DataFrame): power waveform containing two columns "time" and "power", in sec and W respectively.
+        file_prefix (str): prefix for the filename (extension is .MWF by default)
+        file_directory (str): folder to store the mwf file
+        mwf_config (dict): dictionary of instrument control parameters for generating the waveform
 
     Returns:
-        str:
+         path to the maccor waveform file generated
     """
-    # Get file list and validity from json, if ends with .json,
-    # assume it's a file, if not assume it's a json string
-    if file_list_json.endswith(".json"):
-        file_list_data = loadfn(file_list_json)
-    else:
-        file_list_data = json.loads(file_list_json)
 
-    # Setup Events
-    events = KinesisEvents(service='protocolGenerator', mode=file_list_data['mode'])
+    if mwf_config is None:
+        mwf_config = MACCOR_WAVEFORM_CONFIG
 
-    file_list = file_list_data['file_list']
-    all_output_files = []
-    protocol_dir = os.path.join(os.environ.get("BEEP_ROOT", "/"),
-                              processed_dir)
-    for filename in file_list:
-        output_files, result, message = generate_protocol_files_from_csv(
-            filename, output_directory=protocol_dir)
-        all_output_files.extend(output_files)
+    df['power'] = df['power'] / max(abs(df['power'])) * mwf_config['value_scale']
 
-    output_data = {"file_list": all_output_files,
-                   "result": result,
-                   "message": message
-                   }
+    df['step_counter'] = df['power'].diff().fillna(0).ne(0).cumsum()
+    df = df.groupby('step_counter').agg({'time': 'count',
+                                         'power': 'first'})
 
-    events.put_generate_event(output_data, "complete")
+    df.rename(columns={'time': 'duration'}, inplace=True)
 
-    return json.dumps(output_data)
+    df['control_mode'] = mwf_config['control_mode']
 
+    if mwf_config['control_mode'] == 'P':
+        df.loc[df['power'] == 0, 'control_mode'] = 'I'
 
-def main():
-    """Main function for the script"""
-    logger.info('starting', extra=s)
-    logger.info('Running version=%s', __version__, extra=s)
-    try:
-        args = docopt(__doc__)
-        input_json = args['INPUT_JSON']
-        print(process_csv_file_list_from_json(input_json), end="")
-    except Exception as e:
-        logger.error(str(e), extra=s)
-        raise e
-    logger.info('finish', extra=s)
-    return None
+    df['value'] = np.round(abs(df['power']), 5)
 
+    mask = df['power'] <= 0
+    df = df.assign(**{'state': 'C',
+                      'limit_mode': mwf_config['charge_limit_mode'],
+                      'limit_value': mwf_config['charge_limit_value'],
+                      'end_mode': mwf_config['charge_end_mode'],
+                      'operation':  mwf_config['charge_end_operation'],
+                      'end_mode_value': mwf_config['charge_end_mode_value']
+                      })
 
-if __name__ == "__main__":
-    main()
+    df.loc[mask, ['state', 'limit_mode', 'limit_value', 'end_mode', 'operation', 'end_mode_value']] =\
+        ['D', mwf_config['discharge_limit_mode'], mwf_config['discharge_limit_value'],
+         mwf_config['discharge_end_mode'], mwf_config['discharge_end_operation'],
+         mwf_config['discharge_end_mode_value']]
+
+    df['report_mode'] = mwf_config['report_mode']
+    df['report_value'] = mwf_config['report_value']
+    df['range'] = mwf_config['range']
+
+    MWF_file_path = os.path.join(file_directory, file_prefix + ".MWF")
+
+    df[['state', 'control_mode', 'value', 'limit_mode', 'limit_value', 'duration', 'end_mode',
+        'operation', 'end_mode_value', 'report_mode', 'report_value', 'range']].\
+        to_csv(MWF_file_path, sep='\t', header=None, index=None)
+
+    return MWF_file_path
