+++ beep/tests/test_protocol.py
@@ -3,23 +3,20 @@
 
 import os
 import unittest
-import warnings
 import json
-import boto3
 import numpy as np
 import datetime
 
 import pandas as pd
 from beep.utils.secrets_manager import event_setup
 from beep.protocol import PROCEDURE_TEMPLATE_DIR, SCHEDULE_TEMPLATE_DIR
-from beep.generate_protocol import Procedure, \
-    generate_protocol_files_from_csv, convert_velocity_to_power_waveform, generate_maccor_waveform_file
+from beep.generate_protocol import generate_protocol_files_from_csv, convert_velocity_to_power_waveform
+from beep.protocol.maccor import Procedure, generate_maccor_waveform_file
 from beep.protocol.arbin import Schedule
 from beep.protocol.maccor_to_arbin import ProcedureToSchedule
 from monty.tempfile import ScratchDir
 from monty.serialization import dumpfn, loadfn
 from monty.os import makedirs_p
-from botocore.exceptions import NoRegionError, NoCredentialsError
 from beep.utils import os_format, hash_file
 
 import difflib
@@ -39,7 +36,7 @@ class ProcedureTest(unittest.TestCase):
         df_power = convert_velocity_to_power_waveform(velocity_waveform_file, 'mph')
         #Check input and output sizes
         self.assertEqual(len(df_velocity), len(df_power))
-        self.assertTrue(any(df_power['power']<0))
+        self.assertTrue(any(df_power['power'] < 0))
 
     def test_generate_maccor_waveform_file_default(self):
         velocity_waveform_file = os.path.join(TEST_FILE_DIR, "LA4_velocity_waveform.txt")
@@ -48,23 +45,23 @@ class ProcedureTest(unittest.TestCase):
             df_power = convert_velocity_to_power_waveform(velocity_waveform_file, 'mph')
             df_MWF = pd.read_csv(generate_maccor_waveform_file(df_power, "test_LA4_waveform", scratch_dir), sep='\t', header=None)
 
-            #Reference mwf file generated by the cycler for the same power waveform.
+            # Reference mwf file generated by the cycler for the same power waveform.
             df_MWF_ref = pd.read_csv(os.path.join(TEST_FILE_DIR, "LA4_reference_default_settings.mwf"), sep="\t", header=None)
 
             self.assertEqual(df_MWF.shape, df_MWF_ref.shape)
 
-            #Check that the fourth column for charge/discharge limit is empty (default setting)
+            # Check that the fourth column for charge/discharge limit is empty (default setting)
             self.assertTrue(df_MWF.iloc[:,3].isnull().all())
 
-            #Check that sum of durations equals length of the power timeseries
+            # Check that sum of durations equals length of the power timeseries
             self.assertEqual(df_MWF.iloc[:,5].sum(), len(df_power))
 
-            #Check that charge/discharge steps are identical
+            # Check that charge/discharge steps are identical
             self.assertTrue((df_MWF.iloc[:,0] == df_MWF_ref.iloc[:,0]).all())
 
-            #Check that power values are close to each other (col 2)
-            relative_differences = np.abs((df_MWF.iloc[:,2] - df_MWF_ref.iloc[:,2]) / df_MWF_ref.iloc[:,2])
-            self.assertLessEqual(np.mean(relative_differences)*100, 0.01) #mean percentage error < 0.01%
+            # Check that power values are close to each other (col 2)
+            relative_differences = np.abs((df_MWF.iloc[:,2] - df_MWF_ref.iloc[:, 2]) / df_MWF_ref.iloc[:,2])
+            self.assertLessEqual(np.mean(relative_differences)*100, 0.01)  # mean percentage error < 0.01%
 
     def test_generate_maccor_waveform_file_custom(self):
         velocity_waveform_file = os.path.join(TEST_FILE_DIR, "US06_velocity_waveform.txt")
@@ -90,15 +87,15 @@ class ProcedureTest(unittest.TestCase):
                                                                mwf_config=mwf_config), sep='\t', header=None)
             df_MWF_ref = pd.read_csv(os.path.join(TEST_FILE_DIR, "US06_reference_custom_settings.mwf"), sep="\t", header=None)
 
-            #Check dimensions with the reference mwf file
+            # Check dimensions with the reference mwf file
             self.assertEqual(df_MWF.shape, df_MWF_ref.shape)
 
-            #Check that control_mode, charge/discharge state, limit and limit_value columns are identical.
-            self.assertTrue((df_MWF.iloc[:, [0,1,3,4]] == df_MWF_ref.iloc[:, [0,1,3,4]]).all().all())
+            # Check that control_mode, charge/discharge state, limit and limit_value columns are identical.
+            self.assertTrue((df_MWF.iloc[:, [0,1,3,4]] == df_MWF_ref.iloc[:, [0, 1, 3, 4]]).all().all())
 
-            #Check that power values are close to each other (col 2)
-            relative_differences = np.abs((df_MWF.iloc[:,2] - df_MWF_ref.iloc[:,2]) / df_MWF_ref.iloc[:,2])
-            self.assertLessEqual(np.mean(relative_differences)*100, 0.01) #mean percentage error < 0.01%
+            # Check that power values are close to each other (col 2)
+            relative_differences = np.abs((df_MWF.iloc[:,2] - df_MWF_ref.iloc[:, 2]) / df_MWF_ref.iloc[:,2])
+            self.assertLessEqual(np.mean(relative_differences)*100, 0.01)  # mean percentage error < 0.01%
 
 
 class GenerateProcedureTest(unittest.TestCase):
